Scala zio cats

+++ competition-serv-model/src/main/java/compman/compsrv/model/dto/schedule/ScheduleEntryDTO.java
@@ -28,6 +28,6 @@ public class ScheduleEntryDTO {
     private Instant startTime;
     private Instant endTime;
     private Integer numberOfFights;
-    private BigDecimal duration;
+    private int duration;
     private Integer order;
 }
